rename cli yml name
