+++ src/export.ts
@@ -1,7 +1,7 @@
 export * from './ast/export';
 export * from './cairoUtilFuncGen/export';
 export * from './cairoWriter';
-export * from './cli/cli';
+export * from './cli';
 export * from './freeStructWritter';
 export * from './io';
 export * from './nethersolc';
