+++ tests/cli/contract.sol
